# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ontaix API
  version: 1.0.0
  summary: Control-plane API for the Ontaix ontology, proposals, identity, bindings and audit.
  description: |
    The Ontaix API serves the Studio, the admin portal and the MCP gateway.

    Rules that every endpoint follows:

    - Every change to the ontology (concept, specialisation, relation, equivalence, source, binding, attribute, and any rename, deletion, edit or unbind) is a proposal. Ontology mutation endpoints return `202 Accepted` with the proposal; nothing writes an approved ontology row directly.
    - Settings, appearance, identity (groups, members, roles), source enable/disable/reconfigure and agent access are immediate and audited.
    - Authentication is a bearer OIDC ID token or access token. Authorisation is Ontaix-native: roles reach a user only through group membership in this API's own tables.
    - Errors are `application/problem+json` with a stable `code`.
    - Lists that can grow large (concepts, relations, bindings, users, groups, agents, audit) share one pagination, search, filter and sort convention: `page`, `pageSize`, `q`, `filter[<field>]=<value>`, `sort`, `order`, and return `{ items, page, pageSize, total }`.
    - The tenant is resolved from the token; no path carries a tenant id.

...

...

...

...

    CORS: the API answers an explicit origin allow-list from configuration, never a wildcard. Authentication is bearer tokens only, no cookies, so there is no CSRF surface.
  license:
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
  - agentToken: []
...
paths:
...
  /proposals:
    get:
      tags: [Proposals]
      summary: List proposals
      operationId: listProposals
      x-ontaix-permission: model.read
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: A page of proposals, open ones by default (`filter[state]=pending,half_approved`). Filterable - `state`, `type`, `companyId`, `proposerKind`. Sortable - `createdAt`.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Proposal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Proposals]
      summary: Create one proposal from a typed draft
      description: The generic entry used by the teach bar, the MCP gateway and the demo story. Same refusals as the resource endpoints.
      operationId: createProposal
      x-ontaix-permission: proposal.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalDraft'
      responses:
        '429':
          $ref: '#/components/responses/RateLimited'
        '202':
          $ref: '#/components/responses/ProposalAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/Busy'
  /proposals/batch:
    post:
      tags: [Proposals]
      summary: Create several proposals in order
      description: Drafts are created in array order in one transaction so that later drafts can depend on labels introduced by earlier ones. Any refusal fails the whole batch.
      operationId: createProposalBatch
      x-ontaix-permission: proposal.create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drafts]
              properties:
                drafts:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items:
                    $ref: '#/components/schemas/ProposalDraft'
      responses:
        '429':
          $ref: '#/components/responses/RateLimited'
        '202':
          description: The created proposals in order.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Proposal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/Busy'
  /proposals/approve-all:
    post:
      tags: [Proposals]
      summary: Approve every ready proposal, repeatedly
      description: Acts only on proposals the caller may approve under ADR 0003 check points 2 to 4; proposals outside the caller's scope are skipped and counted in `remaining`. Loops while any proposal in scope is ready, at most 200 rounds. Events emitted carry `bulk` set to true.
      operationId: approveAllProposals
      x-ontaix-permission: proposal.approve
      responses:
        '200':
          description: What was approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Busy'
  /proposals/reject-all:
    post:
      tags: [Proposals]
      summary: Reject every open proposal in the caller's scope
      description: Acts only on proposals the caller may reject under ADR 0003 check points 2 to 4; proposals outside the caller's scope are skipped and counted in `remaining`.
      operationId: rejectAllProposals
      x-ontaix-permission: proposal.reject
      responses:
        '200':
          description: What was rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Busy'
  /proposals/finalise-all:
    post:
      tags: [Proposals]
      summary: Play the remaining story and approve everything
      description: Requires `proposal.finalise`, held by Governor at tenant scope, because it creates proposals for every company and approves everything. When `demoStory` is on, plays every remaining scene server-side, then runs approve-all rounds until nothing is ready. When the story is off it equals approve-all. Audited as `demo · finalised all scenes`.
      operationId: finaliseAll
      x-ontaix-permission: proposal.finalise
      responses:
        '200':
          description: The finalise summary used by the caption.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinaliseResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Busy'
  /proposals/{proposalId}:
    parameters:
      - $ref: '#/components/parameters/ProposalId'
    get:
      tags: [Proposals]
      summary: Get one proposal with its readiness
      operationId: getProposal
      x-ontaix-permission: model.read
      responses:
        '200':
          description: The proposal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /proposals/{proposalId}/approve:
    parameters:
      - $ref: '#/components/parameters/ProposalId'
    post:
      tags: [Proposals]
      summary: Approve a proposal
      description: Requires Owner in scope or Governor in an enclosing scope. Refused when not ready (`409 proposal_not_ready`). When `twoApprovers` is on and the type is `change`, the proposal moves to `half_approved` and needs `second-approve` by a different user; calling this endpoint on a `half_approved` proposal behaves as `second-approve`. Otherwise it is applied - pending cleared, domain revision incremented, audit entry written, events emitted. A `rename` or `rename_source` re-checks label uniqueness at approval and is refused with `409 duplicate_label` when the label was taken meanwhile.
      operationId: approveProposal
      x-ontaix-permission: proposal.approve
      responses:
        '200':
          description: The result, including the artefacts that changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Busy'
  /proposals/{proposalId}/second-approve:
    parameters:
      - $ref: '#/components/parameters/ProposalId'
    post:
      tags: [Proposals]
      summary: Give the second approval on a half-approved change
      description: Requires Governor or Owner in scope and a different user from the first approver (`409 same_approver`). Refused unless the state is `half_approved` (`409 proposal_not_half_approved`).
      operationId: secondApproveProposal
      x-ontaix-permission: proposal.second_approve
      responses:
        '200':
          description: The result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Busy'
  /proposals/{proposalId}/reject:
    parameters:
      - $ref: '#/components/parameters/ProposalId'
    post:
      tags: [Proposals]
      summary: Reject a proposal and cascade
      description: Marks the proposal rejected, its concept dying and removes its pending relations, bindings or attribute. Every open proposal whose concept descends from the rejected concept, or whose relation touches it, is rejected in the same transaction and listed in `cascaded`.
      operationId: rejectProposal
      x-ontaix-permission: proposal.reject
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: The result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Busy'

  # ---------------------------------------------------------------- Teach
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC ID token or access token from any provider the tenant trusts (`tenant_identity_provider` maps issuer and audience to the tenant). Authentication only; roles come from Ontaix groups.
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Agent bearer token (client credentials or workload identity). Its issuer and subject must match an `agent` row with `access` on, in a tenant whose `agentAccess` setting is on; otherwise `403`. An agent holds `model.read` and `proposal.create` in its own scope (tenant-wide or one company) and nothing else. Every agent read is logged.
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Rows per page. The portal lists use 40, the agent registry 50.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 40
...
    Filter:
      name: filter
      in: query
      description: Field filters, `filter[field]=value`; a comma separates alternatives. Unknown fields return 400.
      style: deepObject
      explode: true
      schema:
        type: object
        additionalProperties:
          type: string
          maxLength: 200
    Sort:
      name: sort
      in: query
      description: Sort field; each list documents its sortable fields. Unknown fields return 400.
      schema:
        type: string
    Order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
...
    ProposalId:
      name: proposalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ProposalAccepted:
      description: The proposal was created and its pending artefacts exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Proposal'
...
    BadRequest:
      description: Malformed query or body.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The caller's roles do not allow this action in this scope (code `forbidden`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No such resource in this tenant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: The request contradicts the current state; `code` says how.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: The body is well-formed but invalid; `errors` lists the fields.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: The caller's budget for this operation is exhausted (code `rate_limited`); `Retry-After` says when to retry.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Busy:
      description: The tenant's decision lock or a row lock was not granted within the lock timeout, or the transaction hit a deadlock or serialisation failure (code `busy`). Nothing was written; retry the same request after `Retry-After`.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with a stable machine-readable `code`.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: '`urn:ontaix:problem:<code>`'
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - validation_failed
            - duplicate_label
            - duplicate_relation
            - structural_relation
            - cross_company_disabled
            - same_company
            - already_bound
            - source_pending
            - root_concept
            - home_company
            - proposal_not_ready
            - proposal_not_half_approved
            - proposal_decided
            - same_approver
            - locked_setting
            - confirmation_required
            - confirmation_mismatch
            - duplicate_role
            - not_demo_tenant
            - story_finished
            - payload_too_large
            - unsupported_media_type
            - rate_limited
            - duplicate_attribute
            - duplicate_agent
            - unavailable
            - busy
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
    Page:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    Id:
      type: string
      format: uuid
...
    Seed:
      type: number
      minimum: 0
      exclusiveMaximum: 1
      description: Curve bend seed in the range of `relation.seed`. The Studio draws it from its single random stream at draft time so every client draws the same curve; the server stores what it receives and draws one only when the draft carries none.
    DomainKey:
      type: string
      enum: [production, supply, sales, logistics, quality, maintenance, finance, people, engineering]
...
    ActorKind:
      type: string
      enum: [user, agent, system]
    Actor:
      type: object
      required: [kind]
      properties:
        kind:
          $ref: '#/components/schemas/ActorKind'
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
...
    ConceptDraft:
      type: object
      description: The parent is given by id, or by label when it is introduced by an earlier draft of the same batch.
      required: [type, companyId, label, domainKey]
      oneOf:
        - required: [parentId]
        - required: [parentLabel]
      properties:
        type:
          type: string
          const: concept
        companyId:
          $ref: '#/components/schemas/Id'
        parentId:
          $ref: '#/components/schemas/Id'
        parentLabel:
          type: string
          maxLength: 120
          description: Alternative to `parentId` for drafts that depend on a label introduced earlier in the same batch.
        label:
          type: string
          minLength: 1
          maxLength: 120
        domainKey:
          $ref: '#/components/schemas/DomainKey'
        action:
          type: string
          minLength: 1
          maxLength: 60
          default: relates to
          description: Lower-case action of the birth relation.
        reverse:
          type: boolean
          default: false
          description: When true the birth relation reads child action parent.
        caption:
          type: string
          maxLength: 300
        seed:
          $ref: '#/components/schemas/Seed'
    SpecDraft:
      type: object
      description: The parent is given by id, or by label when it is introduced by an earlier draft of the same batch.
      required: [type, companyId, label, domainKey]
      oneOf:
        - required: [parentId]
        - required: [parentLabel]
      properties:
        type:
          type: string
          const: spec
        companyId:
          $ref: '#/components/schemas/Id'
        parentId:
          $ref: '#/components/schemas/Id'
        parentLabel:
          type: string
          maxLength: 120
        label:
          type: string
          minLength: 1
          maxLength: 120
        rule:
          type: string
          maxLength: 200
        domainKey:
          $ref: '#/components/schemas/DomainKey'
        caption:
          type: string
          maxLength: 300
        seed:
          $ref: '#/components/schemas/Seed'
    RelationDraft:
      type: object
      description: Each end is given by id, or by label when it is introduced by an earlier draft of the same batch; `companyId` is required when a label is used.
      required: [type, action]
      allOf:
        - oneOf:
            - required: [aId]
            - required: [aLabel, companyId]
        - oneOf:
            - required: [bId]
            - required: [bLabel, companyId]
      properties:
        type:
          type: string
          const: relation
        companyId:
          $ref: '#/components/schemas/Id'
        aId:
          $ref: '#/components/schemas/Id'
        bId:
          $ref: '#/components/schemas/Id'
        aLabel:
          type: string
          maxLength: 120
        bLabel:
          type: string
          maxLength: 120
        action:
          type: string
          minLength: 1
          maxLength: 60
        caption:
          type: string
          maxLength: 300
        seed:
          $ref: '#/components/schemas/Seed'
...
    SourceDraft:
      type: object
      required: [type, companyId, label, kindText]
      properties:
        type:
          type: string
          const: source
        companyId:
          $ref: '#/components/schemas/Id'
        label:
          type: string
          minLength: 1
          maxLength: 120
        kindText:
          type: string
          maxLength: 60
          description: The catalogue category shown under the label, for example `ERP`.
        connectorCode:
          type: string
          maxLength: 10
        host:
          type: string
          maxLength: 300
        scope:
          type: string
          maxLength: 500
        auth:
          $ref: '#/components/schemas/SourceAuth'
        credentialRef:
          $ref: '#/components/schemas/CredentialRef'
        refresh:
          $ref: '#/components/schemas/RefreshInterval'
        caption:
          type: string
          maxLength: 300
...
    SourceAuth:
      type: string
      enum: [service_principal, oauth2_client_credentials, managed_identity, key_vault_api_key]
...
    BindingDraft:
      type: object
      required: [type, sourceId, conceptIds]
      properties:
        type:
          type: string
          const: bind
        sourceId:
          $ref: '#/components/schemas/Id'
        conceptIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/Id'
        caption:
          type: string
          maxLength: 300
        seed:
          $ref: '#/components/schemas/Seed'
...
    AttributeDraft:
      type: object
      required: [type, conceptId, name, attributeType, col, fill]
      properties:
        type:
          type: string
          const: attr
        conceptId:
          $ref: '#/components/schemas/Id'
        sourceId:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          minLength: 1
          maxLength: 80
        attributeType:
          type: string
          enum: [id, text, number, ref, date]
        col:
          type: string
          maxLength: 200
        fill:
          type: integer
          minimum: 0
          maximum: 100
    ChangeDraft:
      type: object
      description: A change to something already approved. `changeKind` selects which fields of `payload` apply.
      required: [type, changeKind, payload]
      properties:
        type:
          type: string
          const: change
        changeKind:
          type: string
          enum: [rename, delete_concept, edit_relation, remove_relation, unbind, rename_source, remove_source, remove_company, resolve_conflict]
        payload:
          type: object
          properties:
            conceptId:
              $ref: '#/components/schemas/Id'
            newLabel:
              type: string
              minLength: 1
              maxLength: 120
            relationId:
              $ref: '#/components/schemas/Id'
            action:
              type: string
              minLength: 1
              maxLength: 60
            reverse:
              type: boolean
            bindingId:
              $ref: '#/components/schemas/Id'
            sourceId:
              $ref: '#/components/schemas/Id'
            companyId:
              $ref: '#/components/schemas/Id'
            conflictConceptIds:
              type: array
              minItems: 2
              maxItems: 2
              items:
                $ref: '#/components/schemas/Id'
        caption:
          type: string
          maxLength: 300
    ProposalDraft:
      oneOf:
        - $ref: '#/components/schemas/ConceptDraft'
        - $ref: '#/components/schemas/SpecDraft'
        - $ref: '#/components/schemas/RelationDraft'
        - $ref: '#/components/schemas/SourceDraft'
        - $ref: '#/components/schemas/BindingDraft'
        - $ref: '#/components/schemas/AttributeDraft'
        - $ref: '#/components/schemas/ChangeDraft'
      discriminator:
        propertyName: type
        mapping:
          concept: '#/components/schemas/ConceptDraft'
          spec: '#/components/schemas/SpecDraft'
          relation: '#/components/schemas/RelationDraft'
          source: '#/components/schemas/SourceDraft'
          bind: '#/components/schemas/BindingDraft'
          attr: '#/components/schemas/AttributeDraft'
          change: '#/components/schemas/ChangeDraft'
    Proposal:
      type: object
      required: [id, type, state, title, color, html, ready, deps, proposer, createdAt, relationIds, bindingIds, approvals]
      properties:
        id:
          $ref: '#/components/schemas/Id'
        type:
          type: string
          enum: [concept, spec, relation, source, bind, attr, change]
        changeKind:
          type: [string, 'null']
          enum: [rename, delete_concept, edit_relation, remove_relation, unbind, rename_source, remove_source, remove_company, resolve_conflict, remove_cross_company_links, null]
        state:
          type: string
          enum: [pending, half_approved, approved, rejected]
        title:
          type: string
        heading:
          type: string
          description: Panel heading - `New concept`, `Specialisation`, `Relation`, `Change`, `Data source`, `Binding`, `Attribute`, suffixed with ` · <domain name>` when a domain applies.
        color:
          type: string
        companyId:
          type: [string, 'null']
          format: uuid
        domainProductId:
          type: [string, 'null']
          format: uuid
        parentLabel:
          type: [string, 'null']
        deps:
          type: array
          items:
            type: string
          description: Labels and named predicates that must hold before approval.
        ready:
          type: boolean
        waitFor:
          type: [string, 'null']
          description: Shown as `after <waitFor>` while blocked.
        html:
          type: string
          maxLength: 2000
          description: Server-generated; every interpolated label is HTML-escaped and only `<b>`, `<i>` and `<span class="...">` may appear. The Studio renders it with those three tags allow-listed and everything else as text.
        why:
          type: [string, 'null']
        caption:
          type: [string, 'null']
        conceptId:
          type: [string, 'null']
          format: uuid
        relationId:
          type: [string, 'null']
          format: uuid
        relationIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
        sourceId:
          type: [string, 'null']
          format: uuid
        bindingIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
        attributeId:
          type: [string, 'null']
          format: uuid
        proposer:
          $ref: '#/components/schemas/Actor'
        approvals:
          type: array
          maxItems: 2
          items:
            type: object
            required: [ordinal, userId, approvedAt]
            properties:
              ordinal:
                type: integer
                enum: [1, 2]
              userId:
                $ref: '#/components/schemas/Id'
              userName:
                type: string
              approvedAt:
                type: string
                format: date-time
        bulk:
          type: boolean
        createdAt:
          type: string
          format: date-time
        decidedAt:
          type: [string, 'null']
          format: date-time
        artefacts:
          $ref: '#/components/schemas/Artefacts'
    Artefacts:
      type: object
      description: The current state of everything a proposal created or changed, so a client can apply it without a fetch.
      properties:
        concepts:
          type: array
          items:
            $ref: '#/components/schemas/Concept'
        relations:
          type: array
          items:
            $ref: '#/components/schemas/Relation'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Source'
        bindings:
          type: array
          items:
            $ref: '#/components/schemas/Binding'
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/Attribute'
        domainProducts:
          type: array
          items:
            $ref: '#/components/schemas/DomainProduct'
        companies:
          type: array
          items:
            $ref: '#/components/schemas/Company'
    DecisionResult:
      type: object
      required: [proposal, artefacts, cascaded, audit]
      properties:
        proposal:
          $ref: '#/components/schemas/Proposal'
        artefacts:
          $ref: '#/components/schemas/Artefacts'
        cascaded:
          type: array
          description: Proposals rejected by cascade, or created by approval (attribute discovery, conflict detection).
          items:
            $ref: '#/components/schemas/Proposal'
        audit:
          $ref: '#/components/schemas/AuditEntry'
        caption:
          type: string
    BulkResult:
      type: object
      required: [approved, rejected, rounds, remaining]
      properties:
        approved:
          type: integer
        rejected:
          type: integer
        rounds:
          type: integer
        remaining:
          type: integer
          description: Open proposals left, blocked ones included.
        caption:
          type: string
    FinaliseResult:
      type: object
      required: [companies, concepts, bound, equivalences, approved, scenesPlayed]
      properties:
        companies:
          type: integer
        concepts:
          type: integer
        bound:
          type: integer
        equivalences:
          type: integer
        approved:
          type: integer
        scenesPlayed:
          type: integer
        caption:
          type: string
...